requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.3
djangotoolbox==1.8.0
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.9.0
pymongo==2.8
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
whitenoise==3.3.0
